create Navigator panel and reuse it in tabs
