+++ apps/designer/app/designer/features/navigator/index.ts
@@ -1,0 +1,1 @@
+export * from "./navigator";
